Create core Twitter feed functionality
Add and parse JSON into DB
Add Carbon library for time
Display each tweet for each Sub
Modify SubController logic to check for errors and edge cases
Save latest tweets of Subs into DB
Style app with SCSS; design the theme

+++ database/migrations/2016_06_15_162753_create_users_table.php
@@ -13,7 +13,7 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->integer('id')->unique();
+            $table->bigInteger('id')->unique();
             $table->string('handle')->unique();
             $table->string('access_token')->unique();
             $table->rememberToken();
